\name{schoolsAdviceData}
\alias{ps.advice.mat}
\alias{ps.all.vars.mat}
\alias{ps.edge.vars.mat}
\alias{ps.school.vars.mat}
\alias{ps.teacher.vars.mat}


\title{HLSM: Included Data Sets}
\description{
	Data set included with the HLSM package: network variables from Pitts
	and Spillane (2009).
}


\usage{
	ps.advice.mat
	ps.all.vars.mat
	ps.edge.vars.mat
	ps.school.vars.mat
	ps.teacher.vars.mat
}

\format{
	ps.advice.mat: a list of 15 sociomatrices of advice seeking network, one for each school.

	ps.all.vars.mat: a list of 15 arrays of all the covariates, one for each school.
	edge.vars.mat: a list of edge level covariates for 15 different school.
	
	ps.school.vars.mat: a list of school level covariates for all 15 schools.

	ps.teacher.vars.mat: a list of node level covariates for all 15 schools.

	ps.all.vars.mat: a single list of length 15 containing the covariates mentioned above.

 }



\references{
 Pitts, V., & Spillane, J. (2009). "Using social network methods to study school leadership".International Journal of Research & Method in Education, 32, 185-207

 Sweet, T.M., Thomas, A.C., and Junker, B.W. (2012).
 "Hierarchical Network Models for Education Research: Hierarchical Latent Space Models". Journal of Educational and Behavorial Statistics.
}


\author{Sam Adhikari}
\keyword{datasets}


